evolution des foncitons et implementation des mails automatiques

+++ src/main/java/com/example/BookingRoom/Entities/Filiere.java
@@ -20,7 +20,7 @@ public class Filiere {
 
     private String nom;
     private int nombreChambres;
-    private int nombreChambresDisponibles;
+    private double nombreChambresDisponibles;
 
     @ManyToOne
     @JoinColumn(name = "ecole")
